+++ src/test/http_openai_handler_test.cpp
@@ -34,28 +34,6 @@
 
 #include "test_utils.hpp"
 
-class MockedServerRequestInterface final : public tensorflow::serving::net_http::ServerRequestInterface {
-public:
-    MOCK_METHOD(absl::string_view, uri_path, (), (const, override));
-    MOCK_METHOD(absl::string_view, http_method, (), (const, override));
-    MOCK_METHOD(void, WriteResponseBytes, (const char*, int64_t), (override));
-    MOCK_METHOD(void, WriteResponseString, (absl::string_view), (override));
-    MOCK_METHOD((std::unique_ptr<char[], tensorflow::serving::net_http::ServerRequestInterface::BlockDeleter>), ReadRequestBytes, (int64_t*), (override));
-    MOCK_METHOD(absl::string_view, GetRequestHeader, (absl::string_view), (const, override));
-    MOCK_METHOD((std::vector<absl::string_view>), request_headers, (), (const, override));
-    MOCK_METHOD(void, OverwriteResponseHeader, (absl::string_view, absl::string_view), (override));
-    MOCK_METHOD(void, AppendResponseHeader, (absl::string_view, absl::string_view), (override));
-    MOCK_METHOD(void, PartialReplyWithStatus, (tensorflow::serving::net_http::HTTPStatusCode), (override));
-    MOCK_METHOD(void, PartialReply, (std::string), (override));
-    MOCK_METHOD(tensorflow::serving::net_http::ServerRequestInterface::CallbackStatus, PartialReplyWithFlushCallback, ((std::function<void()>)), (override));
-    MOCK_METHOD(tensorflow::serving::net_http::ServerRequestInterface::BodyStatus, response_body_status, (), (override));
-    MOCK_METHOD(tensorflow::serving::net_http::ServerRequestInterface::BodyStatus, request_body_status, (), (override));
-    MOCK_METHOD(void, ReplyWithStatus, (tensorflow::serving::net_http::HTTPStatusCode), (override));
-    MOCK_METHOD(void, Reply, (), (override));
-    MOCK_METHOD(void, Abort, (), (override));
-    MOCK_METHOD(void, PartialReplyEnd, (), (override));
-};
-
 class HttpOpenAIHandlerTest : public ::testing::Test {
 protected:
     ovms::Server& server = ovms::Server::instance();
